refactor(elements): rename utils to styles in Button
